Split main into separate files
