+++ cogs/notifications.py
@@ -2,7 +2,7 @@ from typing import Any
 import discord
 from discord.ext import commands
 from discord.ext.commands import Cog
-from main import Bot, get_config, increment_command_counter, Guild, Notification, OnlineNotification
+from bot import Bot, get_config, increment_command_counter, Guild, Notification, OnlineNotification
 import sys
 sys.path.append('../')
 from datetime import datetime, timedelta, UTC
