+++ cogs/sheets.py
@@ -3,7 +3,7 @@ from discord.ext import commands, tasks
 from discord.ext.commands import Cog
 import sys
 sys.path.append('../')
-from main import Bot, get_config, increment_command_counter
+from bot import Bot, get_config, increment_command_counter
 from datetime import datetime, timedelta, UTC
 import re
 import validators
