+++ scripts/Portables.bat
@@ -1,6 +1,4 @@
-set "params=%*"
-cd %~dp0 && ( if exist "%temp%\getadmin.vbs" del "%temp%\getadmin.vbs" ) && fsutil dirty query %systemdrive% 1>nul 2>nul || (  echo Set UAC = CreateObject^("Shell.Application"^) : UAC.ShellExecute "cmd.exe", "/k cd ""%~sdp0"" && %~s0 %params%", "", "runas", 1 >> "%temp%\getadmin.vbs" && "%temp%\getadmin.vbs" && exit /B )
-cd ..
+cd %~dp0/..
 :loop
 python main.py
 goto loop
